requirements.txt
boto3==1.28.23
fastapi==0.110.1
inference==0.9.20
pydantic==2.6.4
python-dotenv==1.0.1
roboflow==1.1.26
uvicorn==0.29.0
python-multipart==0.0.9
